feat: #185 drawing 기능 hook으로 분리
